# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Street Cred API
  version: "1.0.0"
  description: |
    Server-authoritative backend for Street Cred (Trapcity).

    **Auth**: send `Authorization: Bearer <accessToken>`. Access tokens last 15 minutes; on
    `401 INVALID_TOKEN` call `/auth/refresh` with the stored refresh token and retry once.
    Refresh tokens rotate: always store the newest one.

    **Idempotency**: for actions that spend or grant currency send `Idempotency-Key: <GUID per user action>`.
    A retry with the same key returns the original response instead of charging twice.

    **Rate limits**: any endpoint can answer `429 RATE_LIMITED`; sign-in and sign-up are limited more tightly.

    **Server is authoritative**: apply the `profile` object returned by every mutating call instead of
    changing balances locally.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://staging.example.invalid/api/v1
    description: Staging (to be provisioned)
...
security:
  - bearerAuth: []
paths:
...
  /missions/nearby:
    get:
      tags: [Missions]
      summary: Missions within a radius of a position
      description: "Feeds the map. Returns missions around a position with the distance in metres and, per mission, whether this player can start it right now."
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lng, in: query, required: true, schema: { type: number, minimum: -180, maximum: 180 } }
        - { name: radius, in: query, description: Metres, capped by config (default 10000), schema: { type: number, default: 2000 } }
      responses:
        "200":
          description: Missions sorted by distance
          content:
            application/json:
              schema:
                type: object
                properties:
                  radiusM: { type: number }
                  missions: { type: array, items: { $ref: "#/components/schemas/NearbyMission" } }
        "400": { $ref: "#/components/responses/Error" }
  /missions/history:
    get:
      tags: [Missions]
      summary: Recent attempts by this player
      description: "Recent attempts with their outcome. Use for an activity feed."
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempts: { type: array, items: { $ref: "#/components/schemas/Attempt" } }
...
  /missions/{id}:
    get:
      tags: [Missions]
      summary: One mission
      description: "Details for one mission, e.g. when opening a mission card from a push notification or a saved id."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      responses:
        "200":
          description: Mission
          content:
            application/json:
              schema:
                type: object
                properties: { mission: { $ref: "#/components/schemas/Mission" } }
        "404": { $ref: "#/components/responses/Error" }
  /missions/{id}/attempts:
    post:
      tags: [Missions]
      summary: Start an attempt (charges stamina, fixes the rewards)
      description: "Call when the player confirms \"Start mission\". The server checks distance, level, cooldown and stamina, charges the stamina, and locks in the rewards. Keep the returned attempt id for the rest of the mission."
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Position" }
      responses:
        "201": { $ref: "#/components/responses/AttemptStarted" }
        "403": { $ref: "#/components/responses/Error", description: "TOO_FAR_FROM_MISSION / LEVEL_TOO_LOW" }
        "409": { $ref: "#/components/responses/Error", description: "ATTEMPT_IN_PROGRESS / NOT_ENOUGH_STAMINA / MISSION_ON_COOLDOWN / MISSION_NOT_REPEATABLE" }
  /missions/attempts/{id}/complete:
    post:
      tags: [Missions]
      summary: Finish successfully and receive the reward (once)
      description: "Calling it again returns the same result with `alreadyCompleted: true` and pays nothing extra."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                clientReport:
                  type: object
                  additionalProperties: true
                  description: Optional telemetry (max 4 KB) stored for review; never used to decide rewards.
      responses:
        "404": { $ref: "#/components/responses/Error", description: "No attempt with this id for this player" }
        "200":
          description: Completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt: { $ref: "#/components/schemas/Attempt" }
                  alreadyCompleted: { type: boolean }
                  profile: { $ref: "#/components/schemas/Profile" }
        "409": { $ref: "#/components/responses/Error", description: "ATTEMPT_EXPIRED / ATTEMPT_NOT_ACTIVE" }
        "422": { $ref: "#/components/responses/Error", description: "COMPLETION_TOO_FAST" }
...
  /missions/attempts/{id}/retry:
    post:
      tags: [Missions]
      summary: Free restart after a failed or timed-out attempt
      description: No stamina is charged. Allowed once per failed attempt, within 10 minutes (both configurable).
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Position" }
      responses:
        "404": { $ref: "#/components/responses/Error", description: "No attempt with this id for this player" }
        "201": { $ref: "#/components/responses/AttemptStarted" }
        "409": { $ref: "#/components/responses/Error", description: "NOT_RETRYABLE / ALREADY_RETRIED / ATTEMPT_IN_PROGRESS" }
...
  /garage/{id}/purchase:
    post:
      tags: [Garage]
      summary: Buy a vehicle
      description: "Buy a vehicle from the catalog. The price in cash and gems is taken together; the server refuses if either balance is short."
      parameters:
        - { $ref: "#/components/parameters/CatalogId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      responses:
        "201": { $ref: "#/components/responses/VehicleChanged" }
        "403": { $ref: "#/components/responses/Error", description: "LEVEL_TOO_LOW" }
        "409": { $ref: "#/components/responses/Error", description: "ALREADY_OWNED / INSUFFICIENT_FUNDS" }
  /garage/{id}/equip:
    post:
      tags: [Garage]
      summary: Equip an owned vehicle
      description: "Set the active vehicle. Must be a vehicle the player owns."
      parameters: [{ $ref: "#/components/parameters/CatalogId" }]
      responses:
        "200": { $ref: "#/components/responses/VehicleChanged" }
        "404": { $ref: "#/components/responses/Error", description: "VEHICLE_NOT_OWNED" }
  /garage/{id}/upgrade:
    post:
      tags: [Garage]
      summary: Upgrade one stat by one level
      description: "Cost = upgradeBaseCostCash × (fromLevel + 1). `fromLevel` must match the current level, which makes retries safe."
      parameters:
        - { $ref: "#/components/parameters/CatalogId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stat, fromLevel]
              properties:
                stat: { type: string, enum: [speed, acceleration, handling] }
                fromLevel: { type: integer, minimum: 0 }
      responses:
        "200": { $ref: "#/components/responses/VehicleChanged" }
        "409": { $ref: "#/components/responses/Error", description: "UPGRADE_LEVEL_MISMATCH / MAX_LEVEL_REACHED / INSUFFICIENT_FUNDS" }
  /garage/{id}/customization:
    patch:
      tags: [Garage]
      summary: Paint and underglow
      description: "Save paint and underglow. Send only what changed; anything you leave out keeps its current value."
      parameters:
        - { $ref: "#/components/parameters/CatalogId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                hasCustomPaint: { type: boolean }
                paintColor: { $ref: "#/components/schemas/Color" }
                hasCustomUnderglow: { type: boolean }
                underglowEnabled: { type: boolean }
                underglowColor: { $ref: "#/components/schemas/Color" }
      responses:
        "200": { $ref: "#/components/responses/VehicleChanged" }
        "400": { $ref: "#/components/responses/Error" }
...
  /inventory/{id}/equip:
    post:
      tags: [Inventory & shop]
      summary: Equip an owned outfit
      description: "Wear an owned outfit. The result is reflected in `profile.equippedOutfitId`."
      parameters: [{ $ref: "#/components/parameters/CatalogId" }]
      responses:
        "200":
          description: Equipped
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: "#/components/schemas/InventoryItem" }
                  profile: { $ref: "#/components/schemas/Profile" }
        "404": { $ref: "#/components/responses/Error", description: "ITEM_NOT_OWNED" }
        "409": { $ref: "#/components/responses/Error", description: "NOT_EQUIPPABLE" }
  /inventory/{id}/sell:
    post:
      tags: [Inventory & shop]
      summary: Sell items for cash
      description: "Trade items back for cash at the item's sell value. The last copy of the outfit the player is wearing cannot be sold."
      parameters:
        - { $ref: "#/components/parameters/CatalogId" }
        - { $ref: "#/components/parameters/IdempotencyKeyRequired" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity: { type: integer, minimum: 1, default: 1 }
      responses:
        "200":
          description: Sold
          content:
            application/json:
              schema:
                type: object
                properties:
                  sale:
                    type: object
                    properties:
                      itemId: { type: string }
                      quantitySold: { type: integer }
                      cashEarned: { type: integer }
                      remaining: { type: integer }
                  profile: { $ref: "#/components/schemas/Profile" }
        "409": { $ref: "#/components/responses/Error", description: "NOT_ENOUGH_ITEMS / ITEM_EQUIPPED / NOT_SELLABLE" }
...
  /admin/missions/{id}:
    patch:
      tags: [Admin]
      summary: Edit a mission (only the fields you send; latitude and longitude go together)
      description: "Change a live mission, e.g. its rewards, or set isActive false to take it off the map. Attempts already running keep the rewards they started with."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MissionInput" }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CatalogId:
      name: id
      in: path
      required: true
      description: Catalog id such as `veh_starter_drifter`
      schema: { type: string, pattern: "^[A-Za-z0-9_.-]+$" }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Recommended. A GUID per user action; retries with the same key return the first response.
      schema: { type: string, minLength: 8, maxLength: 128 }
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. A GUID per user action; retries with the same key return the first response.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Error:
      description: Failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    AttemptStarted:
      description: Attempt started
      content:
        application/json:
          schema:
            type: object
            properties:
              attempt: { $ref: "#/components/schemas/Attempt" }
              mission: { $ref: "#/components/schemas/Mission" }
              profile: { $ref: "#/components/schemas/Profile" }
...
    VehicleChanged:
      description: Vehicle updated
      content:
        application/json:
          schema:
            type: object
            properties:
              vehicle: { $ref: "#/components/schemas/OwnedVehicle" }
              profile: { $ref: "#/components/schemas/Profile" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: INSUFFICIENT_FUNDS }
            message: { type: string }
            details: { description: Optional context, e.g. failing fields or availableAt }
...
    Position:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
    Color:
      type: object
      description: Unity SerializableColor, channels 0–1
      required: [r, g, b]
      properties:
        r: { type: number, minimum: 0, maximum: 1 }
        g: { type: number, minimum: 0, maximum: 1 }
        b: { type: number, minimum: 0, maximum: 1 }
        a: { type: number, minimum: 0, maximum: 1, default: 1 }
...
    Stamina:
      type: object
      properties:
        current: { type: integer }
        max: { type: integer }
        regenSeconds: { type: integer }
        nextPointAt: { type: [string, "null"], format: date-time }
        fullAt: { type: [string, "null"], format: date-time }
    Profile:
      type: object
      description: The authoritative player state. Replace local values with this on every response.
      properties:
        level: { type: integer }
        xp: { type: integer, description: Progress inside the current level }
        xpToNextLevel: { type: integer }
        totalXp: { type: integer }
        cash: { type: integer }
        gems: { type: integer }
        stamina: { $ref: "#/components/schemas/Stamina" }
        equippedVehicleId: { type: [string, "null"] }
        equippedOutfitId: { type: [string, "null"] }
        hasFoundersPass: { type: boolean }
...
    Reward:
      type: object
      properties:
        cash: { type: integer }
        gems: { type: integer }
        xp: { type: integer }
        itemId: { type: [string, "null"] }
    Mission:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        missionType: { type: string, enum: [DeadDropCourier, TurfSkirmishAR, ZoneReconWalking, HighRollerSprint] }
        latitude: { type: number }
        longitude: { type: number }
        triggerRadiusM: { type: number }
        timeLimitS: { type: integer }
        requiredStamina: { type: integer }
        reward: { $ref: "#/components/schemas/Reward" }
        minPlayerLevel: { type: integer }
        cooldownS: { type: [integer, "null"], description: null means the mission can only be completed once }
    NearbyMission:
      allOf:
        - { $ref: "#/components/schemas/Mission" }
        - type: object
          properties:
            distanceM: { type: integer }
            availability:
              type: object
              properties:
                canStart: { type: boolean }
                reason: { type: [string, "null"], enum: [LEVEL_TOO_LOW, MISSION_ON_COOLDOWN, MISSION_NOT_REPEATABLE, null] }
                availableAt: { type: [string, "null"], format: date-time }
    Attempt:
      type: object
      properties:
        id: { type: string, format: uuid }
        missionId: { type: string, format: uuid }
        status: { type: string, enum: [active, completed, failed, abandoned, expired] }
        staminaCharged: { type: integer }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
        reward: { $ref: "#/components/schemas/Reward" }
        retryOf: { type: [string, "null"], format: uuid }
        retryCount: { type: integer }
        canRetryFree: { type: boolean }
        result:
          type: [object, "null"]
          properties:
            reward: { $ref: "#/components/schemas/Reward" }
            levelsGained: { type: integer }
            serverElapsedS: { type: integer }
...
    OwnedVehicle:
      type: object
      properties:
        vehicleId: { type: string }
        name: { type: string }
        isEquipped: { type: boolean }
        upgrades:
          type: object
          properties:
            speed: { type: integer }
            acceleration: { type: integer }
            handling: { type: integer }
        maxUpgradeLevel: { type: integer }
        nextUpgradeCostCash:
          type: object
          description: null for a stat already at the maximum level
          properties:
            speed: { type: [integer, "null"] }
            acceleration: { type: [integer, "null"] }
            handling: { type: [integer, "null"] }
        customization:
          type: object
          properties:
            hasCustomPaint: { type: boolean }
            paintColor: { $ref: "#/components/schemas/Color" }
            hasCustomUnderglow: { type: boolean }
            underglowEnabled: { type: boolean }
            underglowColor: { $ref: "#/components/schemas/Color" }
        acquiredAt: { type: string, format: date-time }
...
    InventoryItem:
      type: object
      properties:
        itemId: { type: string }
        name: { type: string }
        category: { type: string }
        rarity: { type: string }
        quantity: { type: integer }
        sellValueCash: { type: integer }
        isEquipped: { type: boolean }
        acquiredAt: { type: string, format: date-time }
...
    MissionInput:
      type: object
      description: POST requires title, missionType, latitude, longitude. PATCH accepts any subset.
      properties:
        title: { type: string }
        description: { type: string }
        missionType: { type: string, enum: [DeadDropCourier, TurfSkirmishAR, ZoneReconWalking, HighRollerSprint] }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        triggerRadiusM: { type: number, default: 30 }
        timeLimitS: { type: integer, default: 120 }
        requiredStamina: { type: integer, default: 15 }
        rewardCash: { type: integer, default: 500 }
        rewardGems: { type: integer, default: 5 }
        rewardXp: { type: integer, default: 250 }
        rewardItemId: { type: [string, "null"] }
        minPlayerLevel: { type: integer, default: 1 }
        cooldownS: { type: [integer, "null"], default: 3600 }
        minCompletionS: { type: integer, default: 10 }
        isActive: { type: boolean, default: true }
        startsAt: { type: [string, "null"], format: date-time }
        endsAt: { type: [string, "null"], format: date-time }